- FPG:     - Implemented MAPRegisterCollection class to encapsulate all MAPRegister management logic: read from FPG, write to FPG, add individual MAPs, remove individual MAPs, find by GraphId function, etc...     - Improved MAP.BaseInfo logic for FPG class.     - Implemented PNG importation for MAPRegister.     - Implemented FPG importation and exportation (not tested yet!). - Implemented DIVFormatBase class to shared common code between DIV format classes. - Fixed opened BinaryReader instances in PAL and MAP classes. - Improved PAL and MAP constructors when import files skiping header and palette data (to acomodate to add to FPG files). - Minor refactorization applied in all project. - Updated test code.

+++ TOOLS/DIV2Tools/DIV2Tools/MethodExtensions/BinaryStreamMethodExtensions.cs
@@ -33,9 +33,9 @@ namespace DIV2Tools.MethodExtensions
         /// </summary>
         /// <param name="stream">This <see cref="BinaryReader"/> instance.</param>
         /// <returns>Returns the <see cref="long"/> value of BaseStream.Length.</returns>
-        public static long GetLength(this BinaryReader stream)
+        public static int GetLength(this BinaryReader stream)
         {
-            return stream.BaseStream.Length;
+            return (int)stream.BaseStream.Length;
         }
 
         /// <summary>
@@ -43,9 +43,9 @@ namespace DIV2Tools.MethodExtensions
         /// </summary>
         /// <param name="stream">This <see cref="BinaryReader"/> instance.</param>
         /// <returns>Returns the current position of reading.</returns>
-        public static long GetCurrentPosition(this BinaryReader stream)
+        public static int GetCurrentPosition(this BinaryReader stream)
         {
-            return stream.BaseStream.Position;
+            return (int)stream.BaseStream.Position;
         }
 
         /// <summary>
@@ -53,9 +53,9 @@ namespace DIV2Tools.MethodExtensions
         /// </summary>
         /// <param name="stream">This <see cref="BinaryWriter"/> instance.</param>
         /// <returns>Returns the current position of writing.</returns>
-        public static long GetCurrentPosition(this BinaryWriter stream)
+        public static int GetCurrentPosition(this BinaryWriter stream)
         {
-            return stream.BaseStream.Position;
+            return (int)stream.BaseStream.Position;
         }
 
         public static bool EOF(this BinaryReader stream)
